headless + hashgrid segv fix

+++ PLSC/include/PLSC/Typedefs.hpp
@@ -5,10 +5,10 @@ typedef bool   b;
 typedef float  f32;
 typedef double f64;
 
-typedef int_fast32_t  i32;
-typedef int_fast64_t  i64;
-typedef uint_fast32_t u32;
-typedef uint_fast64_t u64;
+typedef int  i32;
+typedef long long i64;
+typedef unsigned int u32;
+typedef unsigned long long u64;
 
 typedef int8_t  i8_t;
 typedef int16_t i16_t;
